+++ src/Stack/Package.hs
@@ -35,7 +35,6 @@ import qualified Data.Map.Strict as M
 import qualified Data.Set as S
 import qualified Data.Text as T
 import           Distribution.CabalSpecVersion ( cabalSpecToVersionDigits )
-import qualified Distribution.Compat.NonEmptySet as NES
 import           Distribution.Compiler
                    ( CompilerFlavor (..), PerCompilerFlavor (..) )
 import           Distribution.ModuleName ( ModuleName )
@@ -47,14 +46,12 @@ import           Distribution.PackageDescription
                    , GenericPackageDescription (..), HookedBuildInfo
                    , Library (..), PackageDescription (..), PackageFlag (..)
                    , SetupBuildInfo (..), TestSuite (..), allLibraries
-                   , buildType, depPkgName, depVerRange, maybeToLibraryName
+                   , buildType, depPkgName, depVerRange
                    )
-import           Distribution.Pretty ( prettyShow )
 import           Distribution.Simple.PackageDescription ( readHookedBuildInfo )
 import           Distribution.System ( OS (..), Arch, Platform (..) )
 import           Distribution.Text ( display )
 import qualified Distribution.Types.CondTree as Cabal
-import qualified Distribution.Types.UnqualComponentName as Cabal
 import           Distribution.Utils.Path ( getSymbolicPath )
 import           Distribution.Verbosity ( silent )
 import           Distribution.Version
@@ -96,14 +93,14 @@ import           Stack.Types.Dependency
                    , libraryDepFromVersionRange
                    )
 import           Stack.Types.EnvConfig ( HasEnvConfig )
-import           Stack.Types.GhcPkgId ( ghcPkgIdString )
 import           Stack.Types.NamedComponent
                    ( NamedComponent (..), subLibComponents )
 import           Stack.Types.Package
                    ( BioInput(..), BuildInfoOpts (..), InstallMap
                    , Installed (..), InstalledMap, Package (..)
                    , PackageConfig (..), PackageException (..)
-                   , dotCabalCFilePath, packageIdentifier
+                   , dotCabalCFilePath, installedToPackageIdOpt
+                   , packageIdentifier
                    )
 import           Stack.Types.PackageFile
                    ( DotCabalPath, PackageComponentFile (..) )
@@ -158,7 +155,6 @@ packageFromPackageDescription
           foldAndMakeCollection stackBenchmarkFromCabal $ benchmarks pkg
       , packageExecutables =
           foldAndMakeCollection stackExecutableFromCabal $ executables pkg
-      , packageSubLibDeps = subLibDeps
       , packageBuildType = buildType pkg
       , packageSetupDeps = fmap cabalSetupDepsToStackDep (setupBuildInfo pkg)
       , packageCabalSpec = specVersion pkg
@@ -172,24 +168,6 @@ packageFromPackageDescription
   -- well as use by "stack ghci"
   pkgId = package pkg
   name = pkgName pkgId
-
-  subLibDeps = M.fromList $ concatMap
-    (\(Dependency n vr libs) -> mapMaybe (getSubLibName n vr) (NES.toList libs))
-    (concatMap targetBuildDepends (allBuildInfo' pkg))
-
-  getSubLibName pn vr lib@(LSubLibName _) =
-    Just (MungedPackageName pn lib, libraryDepFromVersionRange vr)
-  getSubLibName _ _ _ = Nothing
-
-toInternalPackageMungedName :: Package -> Text -> Text
-toInternalPackageMungedName pkg =
-    T.pack
-  . prettyShow
-  . MungedPackageName (packageName pkg)
-  . maybeToLibraryName
-  . Just
-  . Cabal.mkUnqualComponentName
-  . T.unpack
 
 -- | This is an action used to collect info needed for "stack ghci". This info
 -- isn't usually needed, so computation of it is deferred.
@@ -218,17 +196,11 @@ getPackageOpts
       let subLibs =
             S.toList $ subLibComponents $ M.keysSet componentsModules
       excludedSubLibs <- mapM (parsePackageNameThrowing . T.unpack) subLibs
-      mungedSubLibs <- mapM
-        ( parsePackageNameThrowing
-        . T.unpack
-        . toInternalPackageMungedName stackPackage
-        )
-        subLibs
       componentsOpts <- generatePkgDescOpts
         installMap
         installedMap
         (excludedSubLibs ++ omitPkgs)
-        (mungedSubLibs ++ addPkgs)
+        addPkgs
         cabalfp
         stackPackage
         componentFiles
@@ -321,8 +293,8 @@ generateBuildInfoOpts BioInput {..} =
   deps =
     concat
       [ case M.lookup name biInstalledMap of
-          Just (_, Stack.Types.Package.Library _ident ipid _) ->
-            ["-package-id=" <> ghcPkgIdString ipid]
+          Just (_, Stack.Types.Package.Library _ident installedInfo) ->
+            installedToPackageIdOpt installedInfo
           _ -> ["-package=" <> packageNameString name <>
             maybe "" -- This empty case applies to e.g. base.
               ((("-" <>) . versionString) . installVersion)