+++ src/Stack/SDist.hs
@@ -70,9 +70,9 @@ import           Stack.Types.EnvConfig
                    ( EnvConfig (..), HasEnvConfig (..), actualCompilerVersionL )
 import           Stack.Types.GhcPkgId ( GhcPkgId )
 import           Stack.Types.Package
-                   ( InstallMap, Installed (..), InstalledMap, LocalPackage (..)
-                   , Package (..), PackageConfig (..), installedVersion
-                   , packageIdentifier
+                   ( InstallMap, Installed (..), InstalledLibraryInfo (..)
+                   , InstalledMap, LocalPackage (..), Package (..)
+                   , PackageConfig (..), installedVersion, packageIdentifier
                    )
 import           Stack.Types.Platform ( HasPlatform (..) )
 import           Stack.Types.PvpBounds ( PvpBounds (..), PvpBoundsType (..) )
@@ -221,8 +221,8 @@ getSDistTarball mpvpBounds pkgDir = do
   (installedMap, _globalDumpPkgs, _snapshotDumpPkgs, _localDumpPkgs) <-
     getInstalled installMap
   let deps = Map.fromList
-        [ (pid, ghcPkgId)
-        | (_, Library pid ghcPkgId _) <- Map.elems installedMap]
+        [ (pid, iliId libInfo)
+        | (_, Library pid libInfo) <- Map.elems installedMap]
   prettyInfoL
     [ flow "Getting the file list for"
     , style File (fromString  pkgFp) <> "."