+++ tests/integration/tests/cabal-sublibrary-dependency/Main.hs
@@ -10,8 +10,4 @@ main :: IO ()
 -- The '--install-ghc' flag is passed here, because IntegrationSpec.runApp sets
 -- up `config.yaml` with `system-ghc: true` and `install-ghc: false`.
 -- (See stack.yaml; using GHC 9.4.7.)
-main = stackErrStderr ["--install-ghc", "build"] $ \str ->
-  let msg = "Sublibrary dependency is not supported, this will almost \
-            \certainly fail."
-  in  unless (msg `isInfixOf` str) $
-        error $ "Expected a warning: \n" ++ show msg
+main = stack ["build", "--install-ghc"]